# operator-api/openapi.yaml
openapi: 3.1.0
info:
  title: Kernel Images API
  version: 0.1.0
paths:
...
  /recording/start:
    post:
      summary: Start a screen recording. Only one recording per ID can be registered at a time.
      operationId: startRecording
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartRecordingRequest"
      responses:
        "201":
          description: Recording started
        "400":
          $ref: "#/components/responses/BadRequestError"
        "409":
          description: A recording is already in progress
          $ref: "#/components/responses/ConflictError"
        "500":
          $ref: "#/components/responses/InternalError"
  /recording/stop:
    post:
      summary: Stop the recording
      operationId: stopRecording
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StopRecordingRequest"
      responses:
        "200":
          description: Recording stopped
        "400":
          $ref: "#/components/responses/BadRequestError"
        "500":
          $ref: "#/components/responses/InternalError"
  /recording/download:
    get:
      summary: Download the most recently recorded video file
      parameters:
        - name: id
          in: query
          description: Optional recorder identifier. When omitted, the server uses the default recorder.
          schema:
            type: string
            pattern: "^[a-zA-Z0-9-]+$"
      operationId: downloadRecording
      responses:
        "200":
          description: Recording file
          headers:
            X-Recording-Started-At:
              description: Timestamp of when the recording started. Guaranteed to be RFC3339.
              schema:
                type: string
            X-Recording-Finished-At:
              description: Timestamp of when the recording finished. Guaranteed to be RFC3339.
              schema:
                type: string
          content:
            video/mp4:
              schema:
                type: string
                format: binary
        "202":
          description: Recording is still in progress, please try again later
          headers:
            Retry-After:
              description: Suggested wait time in seconds before retrying
              schema:
                type: integer
                minimum: 1
        "400":
          $ref: "#/components/responses/BadRequestError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalError"
  /recording/list:
    get:
      summary: List all recorders
      operationId: listRecorders
      responses:
        "200":
          description: List of recorders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RecorderInfo"
        "500":
          $ref: "#/components/responses/InternalError"
  /recording/delete:
    post:
      summary: Delete a previously recorded video file
      operationId: deleteRecording
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteRecordingRequest"
      responses:
        "200":
          description: Recording deleted
        "400":
          $ref: "#/components/responses/BadRequestError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /screenshot/capture:
    post:
      summary: Capture a still image of the desktop
      operationId: captureScreenshot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScreenshotCaptureRequest"
            examples:
              default:
                value:
                  region: { x: 0, y: 0, width: 1280, height: 720 }
                  include_cursor: false
                  display: 0
                  format: png
      responses:
        "200":
          description: Screenshot captured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScreenshotCaptureResponse"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "500":
          $ref: "#/components/responses/InternalError"
  /screenshot/{screenshot_id}:
    get:
      summary: Retrieve captured screenshot by ID
      operationId: getScreenshot
      parameters:
        - in: path
          name: screenshot_id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Image bytes
          content:
            image/png:
              schema: { type: string, format: binary }
            image/jpeg:
              schema: { type: string, format: binary }
        "404":
          $ref: "#/components/responses/NotFoundError"
  /stream/start:
    post:
      summary: Start an RTMPS live stream from the desktop
      operationId: startStream
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StreamStartRequest"
            examples:
              default:
                value:
                  region: null
                  display: 0
                  fps: 30
                  video_codec: h264
                  video_bitrate_kbps: 3500
                  audio: { capture_system: true, capture_mic: false }
                  rtmps_url: "rtmps://live.example.com/app"
                  stream_key: "secret_key"
      responses:
        "200":
          description: Stream starting
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreamStartResponse"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "500":
          $ref: "#/components/responses/InternalError"
  /stream/stop:
    post:
      summary: Stop an RTMPS stream
      operationId: stopStream
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StreamStopRequest"
      responses:
        "200":
          description: Stream stopped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreamStopResponse"
        "404":
          $ref: "#/components/responses/NotFoundError"
...
  /scripts/upload:
    post:
      summary: Upload a script file
      operationId: scriptsUpload
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [path, file]
              properties:
                path: { type: string }
                file: { type: string, format: binary }
                executable: { type: boolean, default: true }
      responses:
        "200":
          description: Uploaded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ScriptUploadResponse" }
        "400":
          $ref: "#/components/responses/BadRequestError"
  /scripts/run:
    post:
      summary: Run an uploaded script (sync or async)
      operationId: scriptsRun
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ScriptRunRequest" }
      responses:
        "200":
          description: Run result or handle
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ProcessExecResponse"
                  - $ref: "#/components/schemas/ScriptRunAsyncResponse"
        "404":
          $ref: "#/components/responses/NotFoundError"
  /scripts/run/{run_id}/logs/stream:
    get:
      summary: Stream logs for an async script run (SSE)
      operationId: scriptsRunLogsStream
      parameters:
        - in: path
          name: run_id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: text/event-stream of stdout/stderr
          headers:
            X-SSE-Content-Type:
              description: Media type of SSE data events (application/json)
              schema:
                type: string
                const: application/json
          content:
            text/event-stream:
              schema:
                type: string
                example: |
                  event: data
                  data: {"stream":"stdout","data_b64":"SGVsbG8K"}
        "404":
          $ref: "#/components/responses/NotFoundError"
...
  /scripts/delete:
    delete:
      summary: Delete an uploaded script
      operationId: scriptsDelete
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ScriptDeleteRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkResponse" }
        "404":
          $ref: "#/components/responses/NotFoundError"
...
components:
  schemas:
...
    StartRecordingRequest:
      type: object
      properties:
        maxFileSizeInMB:
          type: integer
          description: Maximum file size in MB (overrides server default)
          minimum: 10
          maximum: 10000
        framerate:
          type: integer
          description: Recording framerate in fps (overrides server default)
          minimum: 1
          maximum: 20
        maxDurationInSeconds:
          type: integer
          description: Maximum recording duration in seconds (overrides server default)
          minimum: 1
        id:
          type: string
          description: Optional identifier for the recording session. Alphanumeric or hyphen.
          pattern: "^[a-zA-Z0-9-]+$"
      additionalProperties: false
    StopRecordingRequest:
      type: object
      properties:
        forceStop:
          type: boolean
          description: Immediately stop without graceful shutdown. This may result in a corrupted video file.
          default: false
        id:
          type: string
          description: Identifier of the recorder to stop. Alphanumeric or hyphen.
          pattern: "^[a-zA-Z0-9-]+$"
      additionalProperties: false
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
    RecorderInfo:
      type: object
      required: [id, isRecording]
      properties:
        id:
          type: string
        isRecording:
          type: boolean
        started_at:
          type: string
          format: date-time
          nullable: true
          description: Timestamp when recording started
        finished_at:
          type: string
          format: date-time
          nullable: true
          description: Timestamp when recording finished
...
    DeleteRecordingRequest:
      type: object
      properties:
        id:
          type: string
          description: Identifier of the recording to delete. Alphanumeric or hyphen.
          pattern: "^[a-zA-Z0-9-]+$"
      additionalProperties: false
...
    OkResponse:
      type: object
      properties:
        ok: { type: boolean, example: true }
    Region:
      type: object
      properties:
        x: { type: integer, minimum: 0 }
        y: { type: integer, minimum: 0 }
        width: { type: integer, minimum: 1 }
        height: { type: integer, minimum: 1 }
      required: [x, y, width, height]
    ScreenshotCaptureRequest:
      type: object
      properties:
        region: { $ref: "#/components/schemas/Region", nullable: true }
        include_cursor: { type: boolean, default: false }
        display: { type: integer, default: 0 }
        format: { type: string, enum: [png, jpeg], default: png }
        quality: { type: integer, minimum: 1, maximum: 100, nullable: true, description: JPEG only }
      required: [format]
    ScreenshotCaptureResponse:
      type: object
      properties:
        screenshot_id: { type: string, format: uuid }
        content_type: { type: string, example: image/png }
        bytes_b64: { type: string }
    StreamStartRequest:
      type: object
      properties:
        region: { $ref: "#/components/schemas/Region", nullable: true }
        display: { type: integer, default: 0 }
        fps: { type: integer, default: 30 }
        video_codec: { type: string, enum: [h264, h265, av1], default: h264 }
        video_bitrate_kbps: { type: integer, default: 3500 }
        audio:
          type: object
          properties:
            capture_system: { type: boolean, default: true }
            capture_mic: { type: boolean, default: false }
        rtmps_url: { type: string }
        stream_key: { type: string }
      required: [fps, video_codec, rtmps_url, stream_key]
    StreamStartResponse:
      type: object
      properties:
        stream_id: { type: string, format: uuid }
        status: { type: string, example: starting }
        metrics_endpoint: { type: string, example: "/stream/{stream_id}/metrics/stream" }
    StreamStopRequest:
      type: object
      properties:
        stream_id: { type: string, format: uuid }
      required: [stream_id]
    StreamStopResponse:
      type: object
      properties:
        stream_id: { type: string, format: uuid }
        status: { type: string, example: stopped }
...
    ProcessExecResponse:
      type: object
      properties:
        exit_code: { type: integer }
        stdout_b64: { type: string }
        stderr_b64: { type: string }
        duration_ms: { type: integer }
...
    ScriptUploadResponse:
      type: object
      properties:
        path: { type: string }
        size_bytes: { type: integer }
    ScriptRunRequest:
      type: object
      properties:
        path: { type: string }
        args: { type: array, items: { type: string }, default: [] }
        cwd: { type: string, nullable: true }
        env: { type: object, additionalProperties: { type: string }, default: {} }
        as_user: { type: string, nullable: true }
        as_root: { type: boolean, default: false }
        mode: { type: string, enum: [sync, async], default: sync }
        stream: { type: boolean, default: true }
      required: [path]
    ScriptRunAsyncResponse:
      type: object
      properties:
        run_id: { type: string, format: uuid }
...
    ScriptDeleteRequest:
      type: object
      properties:
        path: { type: string }
      required: [path]
...
  responses:
    BadRequestError:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ConflictError:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFoundError:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"